adml_bouncing_ball

Trait FmuFunctions

pub trait FmuFunctions {
    // Required method
    fn do_step(&mut self, current_time: f64, time_step: f64);

    // Provided method
    fn exit_initialization_mode(&mut self) { ... }
}

Functions that must be implemented for any struct that uses this derive macro. The only required one is the do_step. The exit_initialization_mode has default behavior.

Required Methods

fn do_step(&mut self, current_time: f64, time_step: f64)

Steps the simulation forward in time. Necessary for all models.

Provided Methods

fn exit_initialization_mode(&mut self)

Can be used to set internal (private) variables based on parameters from the model description file. It Does nothing by default. Can be overridden if some custom initialization behavior is needed